Request Realtime data instead of Market data
The problem with market data is that it only will return tickers with traded. Indices liek SPX and VIX do not have trades and will never return data on market data. All tickers and indices will retrun data on realtime trades so I have adjusted the request to realtime trades. This will have consequances for the subscription one needs for live trading but one should not trade with delayed data anyway....